Bug that stopped car randomly solved
Before this correction the car was driving fine but without keeping the
line defined by the waypoints exactly. The car used to stop suddenly on
the middle of the road without explanation. After looking the logs of the
DBW_Node I concluded that the problem was on the step before the controller
which is the generation of the points. Then looking at the logs of the
twist_cmd topic I found a dependence between the linear and angular speed.
When the module of the angular speed gets higher, the speed gets reduced,
the problem is that when the angular speed would take negative values, the
linear speed used to get negative values as well making the car braking,
as expected an absolute value was missing on the outputTwist function of
the class PurePursuit within the file pure_pursuit_core.cpp.