DBW node improved
The oscillations caused by the DBW node were minimized  using a PID controller
together with the yaw controller provided by Udacity in order to control
the steering angle, for that the parameters of the PID controller of the
last project were used. Also a low pass filter for the angular velocity is
used, but with a cut frequency 100 times the one of the linear velocity
low pass filter, since the values are 2 orders of magnitude lower.

Also the code of the waypoint follower node was changed so a new angular speed
is calculated on every iteration regarless if the car is following the
trajectory or not. Before this change the last angular speed would be used on
a new cycle if the car was following the trajectory.