Execution frequency of nodes changed
The execution frequency of the nodes tl_detector, twist_controller, waypoint_follower and waypoint_updater
were changed in order to reduce latency. Also the structure of the tl_detector node was changed in order
to allow this. Applying these changes the software seems to execute faster on the simulator when the
camera is activated.

Also a limiter was added to the DBW node in order to avoid big changes on the steering angle. Additional
to the limiter, the cut frequency of the low pass filter for the angular speed was reduced on 1 order of
magnitude in order to avoid big spikes on the angular speed measurement.