Error on styx blocking log messages corrected
The file bridge.py of the node styx was changed because it was trying to
access an inexistent variable of the message package dbw_mkz_msgs making the
roslaunch console always showing a warning which blocked the log messages
of roslog.

NOTE: It can be that this problem was only mine because of my environment,
in that case let me know, we revert this commit and I install the correct
package.