Sudden big impulses on the steering angle corrected
The sudden big impulses on the steering angle were corrected, the problem was
on the steering equation on line 141 of twist_controller.py where the constant
multiplying the output of the pid controller was too big, shooting the
steering angle to the maximum or minimum very easily. Also a limiter for the
target angular speed was added, so it can only change on steps of 0.001
working like a setpoint ramp. The steering control now looks even better than
at the beginning.