Control loss issue when starting a new lap solved
The node waypoint_follower was changed in order to generate uninterrumpted
waypoints when starting a new lap on the simulator, before of this change the
car would drive for a while without waypoints making it loose control at this
point.

Also the same node was changed in order to react to traffic lights that are
more far away than the end of the waypoints. This makes the braking process
softer.

The DBW node was changed again because, though more infrequent, the car would
still lose control at times. With the last change the car drove more than
20 minutes on the simulator without error.