Initial classifier model tested with real images
For that, the traffic light detection node was changed so it selects the right
model depending on the launcher used (styx.launch or site.launch).

In order to test the classifier some debugging code had to be added and other
code had to be commented, since the bag file used didn't publish the pose
which is needed to search for traffic lights. The debugging code is still on
"tl_detector.py" in order to remember where it was, but IT NEEDS TO BE CHANGED
TO THE LAST STATE in order to test further.

Also a small problem was found and corrected on the launch file of the dbw
node for the site general launcher (site.launch).