Changes on DBW node
The parameters of the throttle and brake loops were changed so the program
works well on the v1.2 and v1.3 of the simulator. It is still not working
perfect, P and I of the brake controller may need to be increased and
the throttle loop has big oscillations on the v1.2. It seems that those
oscillations cannot be corrected without breaking the good control on the
version 1.3 of the simulator. Clearly different parameters are needed for
both versions of the simulator.