Speed limit bug fixed, Brake KI = 300
Speed limit bug fixed, it is now automatically taken from waypoint_loader launch file and distributed to all nodes which need the speed limit.

Also the brake KI was increased to 300 to make the vehicle stop before all traffic lights around the test track. Tested for more than one round successfully. Last KI = 15 was much to small to make the vehicle brake early enough.

Todo: Brake Torque in standstill = effectively 700 Nm. This does not yet work and is necessary to work against the creep torque of Carla as mentioned in the info material.